refactor: split order services into smaller specialized services
- Split CartService into CartManagementService and OrderCalculationService
- Split KitchenService into KitchenOperationService and KitchenAnalysisService
- Split OrderService into OrderManagementService
- Each service now handles specific domain functionality
- Improves code organization and maintainability
